back/requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
celery==5.3.4
chardet==5.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==5.2.1
daphne==4.1.2
channels==4.1.0
Django==5.2.5
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
executing==2.2.1
Faker==26.0.0
gunicorn==22.0.0
inflection==0.5.1
ipython==8.26.0
jedi==0.19.2
kombu==5.5.4
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
openpyxl==3.1.5
packaging==25.0
parso==0.8.5
pillow==10.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
reportlab==4.2.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.7.0
